backend/cloud_functions/requirements.txt
firebase-admin>=6.0
firebase-functions>=0.4
google-cloud-firestore>=2.16
